[onert] Move managing cpu tensor into core
This commit moves managing cpu tensor into core
