+++ runtime/onert/backend/cpu/Tensor.h
@@ -17,10 +17,7 @@
 #ifndef __ONERT_BACKEND_CPU_TENSOR_H__
 #define __ONERT_BACKEND_CPU_TENSOR_H__
 
-#include "backend/cpu_common/Allocator.h"
-
-#include <backend/ITensor.h>
-#include <ir/OperandInfo.h>
+#include <backend/cpu_common/Tensor.h>
 
 namespace onert
 {
@@ -29,115 +26,7 @@ namespace backend
 namespace cpu
 {
 
-class Tensor : public ITensor
-{
-public:
-  Tensor() = delete;
-
-public:
-  Tensor(const ir::OperandInfo &info)
-      : _info(info), _buffer(nullptr), _num_references(0), _allocator(nullptr)
-  {
-    // DO NOTHING
-  }
-
-public:
-  // Only one of two method 'setBuffer' must be called once
-  void setBuffer(uint8_t *buffer)
-  {
-    assert(_buffer == nullptr && _allocator == nullptr);
-    _buffer = buffer;
-  }
-  void setBuffer(const std::shared_ptr<cpu_common::Allocator> &alloc)
-  {
-    assert(_buffer == nullptr && _allocator == nullptr);
-    _allocator = alloc;
-  }
-
-  // This works just as setBuffer but it simply overwrite existing Allocator without nullptr check
-  void overwriteBuffer(const std::shared_ptr<cpu_common::Allocator> &alloc) { _allocator = alloc; }
-
-public:
-  uint8_t *buffer() const override
-  {
-    if (_allocator != nullptr)
-      return _allocator->base();
-    else
-      return _buffer;
-  }
-  /**
-   * @brief Get dimension by index
-   *
-   * @param index Index to get diemension
-   * @return size_t Dimension at index
-   * @note N : dimension(0)
-   *       H : dimension(1)
-   *       W : dimension(2)
-   *       C : dimension(3)
-   */
-  size_t dimension(size_t index) const override { return _info.shape().dim(index); }
-  size_t num_dimensions() const override { return _info.shape().rank(); }
-  size_t total_size() const override { return _info.total_size(); }
-  size_t calcOffset(const ir::Coordinates &coords) const override;
-  ir::Layout layout() const override { return ir::Layout::NHWC; }
-  ir::DataType data_type() const override { return _info.typeInfo().type(); }
-  float data_scale() const { return _info.typeInfo().scale(); }
-  int32_t data_offset() const { return _info.typeInfo().offset(); }
-  bool has_padding() const override { return false; }
-  void access(const std::function<void(ITensor &tensor)> &fn) final;
-  bool is_dynamic() const override { return _info.isDynamic(); }
-  void set_dynamic() override { _info.setDynamic(); }
-
-  void increase_ref()
-  {
-    assert(is_dynamic() ||
-           // when not dynamic
-           (_buffer != nullptr || _allocator != nullptr));
-
-    ++_num_references;
-  }
-  void decrease_ref()
-  {
-    assert(_buffer != nullptr || _allocator != nullptr);
-    assert(_num_references > 0);
-    --_num_references;
-    // Only constant tensor has allocator pointer
-    if (_num_references == 0)
-    {
-      if (_buffer != nullptr)
-        _buffer = nullptr;
-      else
-      {
-        _allocator->release();
-        _allocator = nullptr;
-      }
-    }
-  }
-
-  void dimension(size_t index, size_t dim) override
-  {
-    auto rank = _info.shape().rank();
-    rank = rank == 0 ? 1 : rank;
-    if (!(index < static_cast<size_t>(rank)))
-    {
-      throw std::runtime_error("index should be less than rank");
-    }
-
-    _info.shape().dim(index) = dim;
-  }
-
-  void num_dimensions(size_t rank) override
-  {
-    ir::Shape new_shape(rank); // all dims are initialized to 0 (invalid dim)
-    _info.shape(new_shape);
-  };
-
-private:
-  ir::OperandInfo _info;
-  uint8_t *_buffer;
-  int32_t _num_references;
-  std::shared_ptr<cpu_common::Allocator> _allocator;
-};
+using Tensor = cpu_common::Tensor;
 
 } // namespace cpu
 } // namespace backend
